# frontend-handoff/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TraceAI Enterprise Control Plane",
    "version": "1.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AgentRequest": {
        "type": "object",
        "required": ["session_id", "goal"],
        "properties": {
          "session_id": { "type": "string" },
          "goal": { "type": "string" },
          "agent_id": { "type": "string", "nullable": true }
        }
      },
      "AgentResponse": {
        "type": "object",
        "properties": {
          "result": { "type": "string" },
          "request_id": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["queued", "running", "completed", "failed"], "nullable": true }
        }
      }
    }
  },
  "paths": {
...
    "/agent/run": {
      "post": {
        "summary": "Run agent synchronously",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AgentRequest" } } } },
        "responses": { "200": { "description": "Agent response", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AgentResponse" } } } } }
      }
    },
...
    "/traces/{request_id}": {
      "get": {
        "summary": "Get canonical trace",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "request_id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Trace document" } }
      }
    },
...
  }
}